review: one Re-transcribe item and one Missing-word-timing badge
Both were written twice — the menu item byte-identical in the landscape
and portrait Actions menus with its dispatch duplicated in both panels,
and the badge already drifted (the portrait copy had lost the title=
that explains what to do about it).

Both now live in transcript-word-actions.tsx and read the Clips, the
dispatch and the flag out of the editor context themselves, so neither
caller can change one without the other.
