+++ apps/local/app/features/video-editor/components/portrait-studio-panel.tsx
@@ -4,16 +4,16 @@ import { LiveMediaStream } from "./live-media-stream";
 import { SilenceLengthToggle } from "./silence-length-toggle";
 import { RecordingSignalIndicator } from "./timeline-indicators";
 import { StudioActionsDropdown } from "./studio-actions-dropdown";
+import { MissingWordTimingBadge } from "./transcript-word-actions";
 import { PreloadableClipManager } from "../preloadable-clip";
 import {
   getIsOBSActive as getIsOBSActiveSelector,
-  getRetranscribableClipIds,
   getShowCenterLine as getShowCenterLineSelector,
   getShowRecordingSignal as getShowRecordingSignalSelector,
   getShowScrubSlider as getShowScrubSliderSelector,
 } from "../video-editor-selectors";
 import { formatSecondsToTimeCode } from "@/services/utils";
-import { AlertTriangleIcon, SendIcon, VideoOffIcon } from "lucide-react";
+import { SendIcon, VideoOffIcon } from "lucide-react";
 import { useFetcher } from "react-router";
 import { useContextSelector } from "use-context-selector";
 import { VideoEditorContext } from "../video-editor-context";
@@ -68,10 +68,6 @@ export const PortraitStudioPanel = () => {
     (ctx) => ctx.clipsToAggressivelyPreload
   );
   const clips = useContextSelector(VideoEditorContext, (ctx) => ctx.clips);
-  const anyClipsMissingTranscriptWords = useContextSelector(
-    VideoEditorContext,
-    (ctx) => ctx.anyClipsMissingTranscriptWords
-  );
   const clipIdsPreloaded = useContextSelector(
     VideoEditorContext,
     (ctx) => ctx.clipIdsPreloaded
@@ -280,20 +276,9 @@ export const PortraitStudioPanel = () => {
           {videoTitle}
           {" · " + formatSecondsToTimeCode(totalDuration)}
         </span>
-        {anyClipsMissingTranscriptWords && (
-          <span className="text-amber-500 text-xs font-medium inline-flex items-center shrink-0">
-            <AlertTriangleIcon className="size-3.5 mr-1" />
-            Missing word timing
-          </span>
-        )}
+        <MissingWordTimingBadge />
         <div className="flex gap-1 shrink-0">
           <StudioActionsDropdown
-            onRetranscribeAllClips={() =>
-              dispatch({
-                type: "retranscribe-clips",
-                clipIds: getRetranscribableClipIds(clips),
-              })
-            }
             allClipsHaveSilenceDetected={allClipsHaveSilenceDetected}
             allClipsHaveText={allClipsHaveText}
             onExport={() => startExportUpload(videoId, videoTitle)}
